requirements.txt
Django>=5.0
fastapi>=0.110
uvicorn[standard]>=0.29
psycopg2-binary>=2.9
django-environ>=0.11
python-dotenv>=1.0
psutil>=5.9
Pillow>=10.0
django-admin-sortable2>=0.8
pydantic>=2.5
pydantic[email]>=2.5
pyjwt>=2.7
python-multipart>=0.0.6
asgiref>=3.7